Move into new split_screen.rs
